pyproject.toml
[project]
name = "hacks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "django-ninja>=1.3.0",
    "requests>=2.32.3",
]
